refactor: Lamp
